Update README.md
In this task, I created a program for a small business that can help it to manage tasks assigned to each member of the team. This program will work with two text files, user.txt and tasks.txt. tasks.txt stores a list of all the tasks that the team is working on. The data for each task is stored on a separate line in the text file.  user.txt stores user names and passwords